refactor refs
